+++ Intwenty/Entity/ExportModel.cs
@@ -17,7 +17,7 @@ namespace Intwenty.Entity
 
         public List<DatabaseItem> DatabaseItems { get; set; }
 
-        public List<UserInterfaceItem> UserInterfaceItems { get; set; }
+        public List<UserInterfaceStructureItem> UserInterfaceItems { get; set; }
 
         public List<DataViewItem> DataViewItems { get; set; }
 
@@ -33,7 +33,7 @@ namespace Intwenty.Entity
             ModelDate = DateTime.Now;
             Applications = new List<ApplicationItem>();
             DatabaseItems = new List<DatabaseItem>();
-            UserInterfaceItems = new List<UserInterfaceItem>();
+            UserInterfaceItems = new List<UserInterfaceStructureItem>();
             DataViewItems = new List<DataViewItem>();
             Translations = new List<TranslationItem>();
             ValueDomains = new List<ValueDomainItem>();
