+++ Intwenty/Model/IntwentyRegistry.cs
@@ -29,34 +29,35 @@ namespace Intwenty.Model
             _metatypes = new List<IntwentyMetaType>();
 
             _metatypes.Add(new IntwentyMetaType() { Code = ApplicationModelItem.MetaTypeApplication, Title = "Application", ModelCode = "APPMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeCheckBox, Title = "Checkbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeComboBox, Title = "Combobox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeUserInterface, Title = "User Interface" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeDatePicker, Title = "Datepicker" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGrid, Title = "Grid" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridCheckBox, Title = "Grid Checkbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridComboBox, Title = "Grid Combobox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridDatePicker, Title = "Grid Datepicker" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridEmailBox, Title = "Grid Emailbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridLookUp, Title = "Grid Look Up" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridNumBox, Title = "Grid Numbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridStaticHTML, Title = "Grid HTML" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEditGridTextBox, Title = "Grid Textbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeEmailBox, Title = "Emailbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeImage, Title = "Image" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeImageBox, Title = "Image Upload" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeLabel, Title = "Label" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeTextListColumn, Title = "Text List Column" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeLookUp, Title = "Look Up" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeNumBox, Title = "Numbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypePanel, Title = "Panel" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypePasswordBox, Title = "Password Box" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeSection, Title = "Section" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeStaticHTML, Title = "HTML" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeTextArea, Title = "Text Area" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeTextBlock, Title = "Text Block" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeTextBox, Title = "Textbox" , ModelCode = "UIMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeMultiSelect, Title = "Multi Select", ModelCode = "UIMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeInputUI, Title = "Input user interface", ModelCode = "UMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceModelItem.MetaTypeListUI, Title = "List user interface", ModelCode = "UIMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeCheckBox, Title = "Checkbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeComboBox, Title = "Combobox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeDatePicker, Title = "Datepicker" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGrid, Title = "Grid" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridCheckBox, Title = "Grid Checkbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridComboBox, Title = "Grid Combobox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridDatePicker, Title = "Grid Datepicker" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridEmailBox, Title = "Grid Emailbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridLookUp, Title = "Grid Look Up" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridNumBox, Title = "Grid Numbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridStaticHTML, Title = "Grid HTML" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEditGridTextBox, Title = "Grid Textbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeEmailBox, Title = "Emailbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeImage, Title = "Image" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeImageBox, Title = "Image Upload" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeLabel, Title = "Label" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeTextListColumn, Title = "Text List Column" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeLookUp, Title = "Look Up" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeNumBox, Title = "Numbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypePanel, Title = "Panel" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypePasswordBox, Title = "Password Box" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeSection, Title = "Section" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeStaticHTML, Title = "HTML" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeTextArea, Title = "Text Area" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeTextBlock, Title = "Text Block" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeTextBox, Title = "Textbox" , ModelCode = "UISTRUCTUREMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = UserInterfaceStructureModelItem.MetaTypeMultiSelect, Title = "Multi Select", ModelCode = "UISTRUCTUREMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = DatabaseModelItem.MetaTypeDataColumn, Title = "Column", ModelCode = "DATAMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = DatabaseModelItem.MetaTypeDataTable, Title = "Table", ModelCode = "DATAMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = DataViewModelItem.MetaTypeDataView, Title = "Data View", ModelCode = "DATAVIEWMODEL" });
@@ -68,8 +69,7 @@ namespace Intwenty.Model
             _metatypes.Add(new IntwentyMetaType() { Code = EndpointModelItem.MetaTypeDataViewList, Title = "View (POST)", ModelCode = "ENDPOINTMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = EndpointModelItem.MetaTypeCustomPost, Title = "Custom (POST)", ModelCode = "ENDPOINTMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = EndpointModelItem.MetaTypeCustomGet, Title = "Custom (GET)", ModelCode = "ENDPOINTMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = ViewModelItem.MetaTypeInputView, Title = "Input View", ModelCode = "UIVIEWMODEL" });
-            _metatypes.Add(new IntwentyMetaType() { Code = ViewModelItem.MetaTypeListView, Title = "List View", ModelCode = "UIVIEWMODEL" });
+            _metatypes.Add(new IntwentyMetaType() { Code = ViewModelItem.MetaTypeUIView, Title = "Input View", ModelCode = "UIVIEWMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = FunctionModelItem.MetaTypeCreate, Title = "Create", ModelCode = "FUNCTIONMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = FunctionModelItem.MetaTypeSave, Title = "Save", ModelCode = "FUNCTIONMODEL" });
             _metatypes.Add(new IntwentyMetaType() { Code = FunctionModelItem.MetaTypeDelete, Title = "Delete", ModelCode = "FUNCTIONMODEL" });
@@ -123,56 +123,56 @@ namespace Intwenty.Model
 
 
             prop = new IntwentyProperty("COLLAPSIBLE", "Collapsible", "BOOLEAN");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeSection);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeSection);
             _properties.Add(prop);
 
             prop = new IntwentyProperty("STARTEXPANDED", "Start expanded", "BOOLEAN");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeSection);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeSection);
             _properties.Add(prop);
 
             prop = new IntwentyProperty("READONLY", "Read Only", "BOOLEAN");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeCheckBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeComboBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeDatePicker);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGridCheckBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGridComboBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGridDatePicker);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGridEmailBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGridLookUp);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGridTextBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEmailBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeLookUp);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeNumBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypePasswordBox);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeTextArea);
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeTextBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeCheckBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeComboBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeDatePicker);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGridCheckBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGridComboBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGridDatePicker);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGridEmailBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGridLookUp);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGridTextBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEmailBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeLookUp);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeNumBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypePasswordBox);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeTextArea);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeTextBox);
             _properties.Add(prop);
 
 
             prop = new IntwentyProperty("EDITMODE", "Edit Mode", "LIST");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGrid);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGrid);
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "CELL", DisplayValue = "Line" });
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "MODAL", DisplayValue = "Modal" });
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "EXPANDER", DisplayValue = "Expander" });
             _properties.Add(prop);
 
             prop = new IntwentyProperty("GRIDLAYOUT", "Layout", "LIST");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeEditGrid);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeEditGrid);
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "GRID", DisplayValue = "Grid" });
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "CARD", DisplayValue = "Card" });
             _properties.Add(prop);
 
             prop = new IntwentyProperty("IMGWIDTH", "Width", "NUMERIC");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeImage);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeImage);
             _properties.Add(prop);
 
             prop = new IntwentyProperty("IMGLAYOUT", "Layout", "LIST");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeImage);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeImage);
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "DEFAULT", DisplayValue = "Default" });
             _properties.Add(prop);
 
             prop = new IntwentyProperty("LABELSIZE", "Size", "LIST");
-            prop.ValidFor.Add(UserInterfaceModelItem.MetaTypeLabel);
+            prop.ValidFor.Add(UserInterfaceStructureModelItem.MetaTypeLabel);
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "H1", DisplayValue = "h1" });
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "H2", DisplayValue = "h2" });
             prop.ValidValues.Add(new SelectablePropertyValue() { CodeValue = "H3", DisplayValue = "h3" });
