+++ Intwenty/Views/Application/Create.cshtml
@@ -2,7 +2,8 @@
 
 @{
     var title = "";
-    var view = Model.UIViews.Find(p => p.IsMetaTypeCreateView);
+    //var view = Model.UIViews.Find(p => p.IsMetaTypeCreateView);
+    Intwenty.Model.UIRendering.UIView view = null;
     if (view != null)
     {
         title = view.LocalizedTitle;
