+++ Intwenty/Views/Application/Detail.cshtml
@@ -3,12 +3,13 @@
 
 @{
     var title = "";
-    var view = Model.UIViews.Find(p => p.IsMetaTypeDetailView);
+    //var view = Model.UIViews.Find(p => p.IsMetaTypeDetailView);
+    Intwenty.Model.UIRendering.UIView view = null;
     if (view != null)
     {
         title = view.LocalizedTitle;
     }
-    
+
 
     ViewData["Title"] = title;
