+++ Intwenty/Views/Application/EditList.cshtml
@@ -11,7 +11,8 @@
         @{
             var pagesize = 20;
             var title = "";
-            var view = Model.UIViews.Find(p => p.IsMetaTypeEditListView);
+            //var view = Model.UIViews.Find(p => p.IsMetaTypeEditListView);
+            Intwenty.Model.UIRendering.UIView view = null;
             if (view != null)
             {
                 title = view.LocalizedTitle;
