+++ Intwenty/Views/Shared/UIControls/_CheckBox.cshtml
@@ -5,7 +5,7 @@
            id="@Model.UIId"
            name="@Model.UIId"
            v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName"
-           data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeCheckBox
+           data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeCheckBox
            data-title="@Model.Title"
            data-dbtable="@Model.DataTableDbName" 
            data-dbfield="@Model.DataColumnDbName"
