+++ Intwenty/Views/Shared/UIControls/_ComboBox.cshtml
@@ -15,7 +15,7 @@
                 v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName"
                 class="form-control form-control-sm"
                 v-on:change="onUserInput(event)"
-                data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeComboBox
+                data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeComboBox
                 data-validationfield="@validationfield"
                 data-title="@Model.Title"
                 data-dbtable="@Model.DataTableDbName"
