+++ Intwenty/Views/Shared/UIControls/_DatePicker.cshtml
@@ -17,7 +17,7 @@
                class="form-control form-control-sm"
                v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName"
                v-on:input="onUserInput(event)"
-               data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeDatePicker
+               data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeDatePicker
                data-title="@Model.Title"
                data-validationfield="@validationfield"
                data-dbtable="@Model.DataTableDbName"
