+++ Intwenty/Views/Shared/UIControls/_EmailBox.cshtml
@@ -15,7 +15,7 @@
                class="form-control form-control-sm"
                v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName"
                v-on:input="onUserInput(event)"
-               data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeEmailBox
+               data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeEmailBox
                data-title="@Model.Title"
                data-validationfield="@validationfield"
                data-dbtable="@Model.DataTableDbName"
