+++ Intwenty/Views/Shared/UIControls/_ImageBox.cshtml
@@ -14,7 +14,7 @@
                    name="@Model.UIId"
                    type="file"
                    class="uploadInput"
-                   data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeImageBox
+                   data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeImageBox
                    data-title="@Model.LocalizedTitle"
                    data-validationfield="@validationfield"
                    data-dbtable="@Model.DataTableDbName"
