+++ Intwenty/Views/Shared/UIControls/_LookUp.cshtml
@@ -19,7 +19,7 @@
                class="form-control form-control-sm" 
                v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName" 
                v-on:keydown.tab="getDataViewValue('@Model.DataViewMetaCode','@Model.DataColumnDbName','@Model.UIId')" 
-               data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeLookUp
+               data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeLookUp
                data-title="@Model.LocalizedTitle"
                data-dbtable="@Model.DataTableDbName" 
                data-dbfield="@Model.DataColumnDbName"
