+++ Intwenty/Views/Shared/UIControls/_MultiSelect.cshtml
@@ -14,7 +14,7 @@
                  class="form-control form-control-lg" 
                  v-bind:idfield.sync="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName" 
                  v-bind:textfield.sync="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumn2DbName"
-                 data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeMultiSelect
+                 data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeMultiSelect
                  data-title="@Model.LocalizedTitle"
                  data-validationfield="@validationfield"
                  data-dbtable="@Model.DataTableDbName"
