+++ Intwenty/Views/Shared/UIControls/_NumBox.cshtml
@@ -11,7 +11,7 @@
            class="form-control form-control-sm" 
            v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName" 
            v-on:input="onUserInput(event)" 
-           data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeNumBox
+           data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeNumBox
            data-title="@Model.LocalizedTitle"
            data-dbtable="@Model.DataTableDbName" 
            data-dbfield="@Model.DataColumnDbName"
