+++ Intwenty/Views/Shared/UIControls/_PasswordBox.cshtml
@@ -15,7 +15,7 @@
            class="form-control form-control-sm"
            v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName"
            v-on:input="onUserInput(event)"
-           data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypePasswordBox
+           data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypePasswordBox
            data-title="@Model.LocalizedTitle"
            data-validationfield="@validationfield"
            data-dbtable="@Model.DataTableDbName"
