+++ Intwenty/Views/Shared/UIControls/_TextArea.cshtml
@@ -12,7 +12,7 @@
               cols="50" 
               class="form-control form-control-sm" 
               v-on:input="onUserInput(event)" 
-              data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeTextArea
+              data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeTextArea
               data-title="@Model.LocalizedTitle"
               data-dbtable="@Model.DataTableDbName" 
               data-dbfield="@Model.DataColumnDbName"
