+++ Intwenty/Views/Shared/UIControls/_TextBox.cshtml
@@ -17,7 +17,7 @@
                class="form-control form-control-sm"
                v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName"
                v-on:input="onUserInput(event)"
-               data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeTextBox
+               data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeTextBox
                data-title="@Model.LocalizedTitle"
                data-validationfield="@validationfield"
                data-dbtable="@Model.DataTableDbName"
@@ -33,7 +33,7 @@
                type="text"
                class="form-control form-control-sm"
                v-model="@Model.JavaScriptObjectName.@Model.DataTableDbName.@Model.DataColumnDbName"
-               data-metatype=@Intwenty.Model.UserInterfaceModelItem.MetaTypeTextBox
+               data-metatype=@Intwenty.Model.UserInterfaceStructureModelItem.MetaTypeTextBox
                data-title="@Model.LocalizedTitle"
                disabled />
